playwright-python-automation/requirements.txt
pytest-playwright
playwright==1.48.0
pytest==8.3.2
python-dotenv==1.0.1
pytest-html==4.1.1
pytest-xdist==3.6.1
json-diff
Faker



